apps/ai-service/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
numpy
opencv-python
pillow
jinja2
scipy
scikit-learn
requests
